backend/python/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.24.3
nltk>=3.8.1
scikit-learn>=1.3.2
python-dotenv>=1.0.0

# AWS and LangChain dependencies
boto3>=1.34.72
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.10

# Vector store and embeddings
faiss-cpu>=1.7.4

# Optional: Keep sentence-transformers as fallback
sentence-transformers>=2.2.2